app→CJS module, tests are ESM.
Was getting the following error[1] when running the tests:

```
Exception during run: Error [ERR_REQUIRE_ESM]: require() of ES Module
```

It looks like chai is now a ESM, which must be brought in via `import`
instead of `require`. Could not switch entire app over to ESM as it
appears the `express` package is still CJS. Rename the files to
specify these two scopes.

There is likely a cleaner way to fix this.

[1] https://github.com/cmumford/node-starter/actions/runs/13972326884/job/39117094238#step:5:25